requirements.txt
# Blocket scraper dependencies
httpx>=0.27.0
asyncio>=3.4.3

# Database
psycopg2-binary>=2.9.9

# Optional: for loading .env files
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0  # For parallel test execution
